Move legacy files to separate folder
