+++ legacy.html
@@ -13,11 +13,11 @@
 
         <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.4.1/jquery.min.js"></script>
 
-        <script src="lib/js/videoproof/user-guide.js" defer></script>
+        <script src="lib/js/legacy/user-guide.js" defer></script>
 
         <link rel="stylesheet" href="lib/css/videoproof.css">
         <link rel="stylesheet" href="lib/css/videoproof/form-controls.css">
-        <script src="lib/js/main.mjs" async type="module"></script>
+        <script src="lib/js/legacy/main.mjs" async type="module"></script>
 
 <script>
     // With this pattern it's save to load the controller javascript asyncronously.
