+++ lib/js/legacy/content-editable.mjs
