+++ lib/js/legacy/layouts/vartools-typespec.mjs
