+++ lib/js/legacy/layouts/videoproof-array.mjs
