+++ lib/js/legacy/layouts/videoproof-contextual.mjs
@@ -1,6 +1,6 @@
 /* jshint esversion: 11, browser: true, unused:true, undef:true, laxcomma: true, laxbreak: true */
 
-import DOMTool from '../domTool.mjs';
+import DOMTool from '../../domTool.mjs';
 
 function _testCharType(extendedCharGroups, c, re) {
     if (re.test(c))
