+++ lib/js/legacy/layouts/videoproof-type-your-own.mjs
