+++ lib/js/legacy/text-selection.mjs
