+++ lib/js/legacy/user-guide.js
