+++ lib/js/legacy/videoproof-controller.mjs
@@ -1,14 +1,14 @@
 /* jshint esversion: 11, browser: true, unused:true, undef:true, laxcomma: true, laxbreak: true, devel: true */
 /*eslint no-prototype-builtins: "off", no-self-assign: "off"*/
 
-import * as opentype from './vendor/opentype.js/dist/opentype.mjs';
+import * as opentype from '../vendor/opentype.js/dist/opentype.mjs';
 import {  normalizeRanges, mergePathRanges, getFullPathsFromRanges
        , normalizePathsRanges, serializePathRanges, deserializePathRanges
        , markupSelectionInline, markupSelectionStructureSave
        , clipAndFilterRanges
        } from './text-selection.mjs';
 
-import {getDecompressSync} from './vendor/wawoff2/decompress.mjs';
+import {getDecompressSync} from '../vendor/wawoff2/decompress.mjs';
 const woff2decompress = await getDecompressSync();
 
 import {init as initGrid} from './layouts/videoproof-array.mjs';
@@ -20,13 +20,13 @@ import {init as initTypespec
 import {init as initWaterfall} from './layouts/vartools-waterfall.mjs';
 import {init as initContextual} from './layouts/videoproof-contextual.mjs';
 import {init as initVarToolsGrid} from './layouts/vartools-grid.mjs';
-import DOMTool from './domTool.mjs';
-import LocalFontStorage from './local-font-storage.mjs';
-import {OTFeatureInfo} from './ot-feature-info.mjs';
+import DOMTool from '../domTool.mjs';
+import LocalFontStorage from '../local-font-storage.mjs';
+import {OTFeatureInfo} from '../ot-feature-info.mjs';
 import {
     fixGridLineBreaks
   , fixContextualLineBreaks
-} from './affixed-line-breaks.mjs';
+} from '../affixed-line-breaks.mjs';
 
 /**
  * Array.from( cartesianProductGen([['a', 'b'], ['c', 'd']]) )
